+++ lib/src/components/carousel_widget_3d.dart
@@ -7,6 +7,8 @@ import 'package:flutter_3d_carousel/src/utils/util_functions.dart';
 
 part 'sub_widget.dart';
 
+// TODO: Create Presets: .horizontal, .vertical, .spinning
+
 /// A 3D carousel widget that displays children in a circular arrangement with
 /// 3D perspective effects and interactive rotation.
 ///
@@ -54,24 +56,29 @@ class CarouselWidget3D extends StatefulWidget {
   /// Whether it is a horizontal carousel or a vertical carousel.
   ///
   /// Defaults to [Axis.horizontal]
-  final Axis spinAxis;
+  final Axis dragDirection;
 
   /// The amount of blur effect applied to background children.
   ///
   /// Defaults to 3.0
   final double backgroundBlur;
 
-  /// Whether the children of the carousel should spin with the carousel.
+  /// TODO: Whether the children of the carousel should spin with the carousel.
   /// If true, the children of the carousel will always be facing outwards from the center of the carousel.
   /// If false, the children will always be facing forward (facing the camera or viewer).
   ///
   /// Defaults to true
-  final bool spinWhileRotating;
+  final bool childrenAreAlwaysFacingForward;
 
   /// Whether the carousel should automatically rotate.
   ///
   /// Defaults to true
-  final bool shouldRotate;
+  final bool thetaShouldRotate;
+
+  /// Whether the carousel should automatically rotate.
+  ///
+  /// Defaults to true
+  final bool alphaShouldRotate;
 
   /// The time in milliseconds for a complete 360-degree revolution when auto-rotating.
   ///
@@ -93,14 +100,21 @@ class CarouselWidget3D extends StatefulWidget {
   /// Defaults to 1.0
   final double dragSensitivity;
 
+  /// TODO:
+  final TiltController tiltController;
+
   /// Determines how the edges of the carousel area will be clipped.
   ///
   /// Defaults to [Clip.hardEdge].
   final Clip clipBehavior;
 
-  /// Called whenever the internal AnimationController's value is updated with the new value.
+  /// Called whenever the internal AnimationController value for Theta is updated with the new value.
   /// The internal AnimationController animates from 0 to 2 * pi.
-  final Function(double)? onValueChanged;
+  final Function(double)? onThetaValueChanged;
+
+  /// Called whenever the internal AnimationController value for Alpha is updated with the new value.
+  /// The internal AnimationController animates from 0 to 2 * pi.
+  final Function(double)? onAlphaValueChanged;
 
   /// A widget to fill the background with.
   ///
@@ -129,16 +143,19 @@ class CarouselWidget3D extends StatefulWidget {
     this.isDragInteractive = true,
     this.onlyRenderForeground = false,
     this.clockwise = false,
-    this.spinAxis = Axis.horizontal,
+    this.dragDirection = Axis.horizontal,
     this.backgroundBlur = 3.0,
-    this.spinWhileRotating = true,
-    this.shouldRotate = true,
+    this.childrenAreAlwaysFacingForward = false,
+    this.thetaShouldRotate = true,
+    this.alphaShouldRotate = false,
     this.timeForFullRevolution = 20000,
     this.snapTimeInMillis = 100,
     this.perspectiveStrength = 0.001,
     this.dragSensitivity = 1.0,
+    this.tiltController = const TiltController(initialTiltAngle: 0),
     this.clipBehavior = Clip.hardEdge,
-    this.onValueChanged,
+    this.onThetaValueChanged,
+    this.onAlphaValueChanged,
     this.background,
     this.core,
     required this.children,
@@ -148,10 +165,14 @@ class CarouselWidget3D extends StatefulWidget {
   State<CarouselWidget3D> createState() => _CarouselWidget3DState();
 }
 
-class _CarouselWidget3DState extends State<CarouselWidget3D>
-    with SingleTickerProviderStateMixin {
-  late final AnimationController _controller;
+// TODO: Handle all Gesture Responses
 
+class _CarouselWidget3DState extends State<CarouselWidget3D>
+    with TickerProviderStateMixin {
+  late final AnimationController thetaController;
+  late final AnimationController alphaController;
+
+  // The theta angle between successive elements in the carousel
   late final double _stepAngle;
 
   static const int animationTimeMillis = 300;
@@ -161,40 +182,65 @@ class _CarouselWidget3DState extends State<CarouselWidget3D>
     super.initState();
     _stepAngle = (2 * math.pi) / widget.children.length;
 
-    _controller = AnimationController(
+    thetaController = AnimationController(
       vsync: this,
       value: 0.0,
       lowerBound: 0.0,
       upperBound: 2 * math.pi,
       duration: Duration(milliseconds: widget.timeForFullRevolution),
     );
-    _controller.addListener(_controllerListener);
-    if (widget.shouldRotate) {
-      _rotateInfinitely();
+    thetaController.addListener(_thetaControllerListener);
+    if (widget.thetaShouldRotate) {
+      thetaController.repeat();
+    }
+
+    alphaController = AnimationController(
+      vsync: this,
+      value: widget.tiltController.initialTiltAngle,
+      lowerBound: 0.0,
+      upperBound: 2 * math.pi,
+      duration: Duration(milliseconds: widget.tiltController.tiltAnimationTime),
+    );
+    alphaController.addListener(_alphaControllerListener);
+    if (widget.alphaShouldRotate) {
+      alphaController.repeat();
     }
   }
 
-  bool isAnimating = false;
+  bool thetaIsAnimating = false;
+  bool alphaIsAnimating = false;
 
   @override
   void dispose() {
-    _controller.removeListener(_controllerListener);
-    _controller.dispose();
+    thetaController.removeListener(_thetaControllerListener);
+    thetaController.dispose();
+
+    alphaController.removeListener(_alphaControllerListener);
+    alphaController.dispose();
     super.dispose();
   }
 
-  void _controllerListener() {
-    if (_controller.isAnimating) {
-      isAnimating = true;
+  void _thetaControllerListener() {
+    if (thetaController.isAnimating) {
+      thetaIsAnimating = true;
     } else {
-      isAnimating = false;
+      thetaIsAnimating = false;
     }
-    if (widget.onValueChanged != null) {
-      widget.onValueChanged!(_controller.value);
+    if (widget.onThetaValueChanged != null) {
+      widget.onThetaValueChanged!(thetaController.value);
     }
   }
 
-  void _rotateInfinitely() => _controller.repeat();
+  void _alphaControllerListener() {
+    if (alphaController.isAnimating) {
+      alphaIsAnimating = true;
+    } else {
+      alphaIsAnimating = false;
+    }
+    if (widget.onAlphaValueChanged != null) {
+      widget.onAlphaValueChanged!(alphaController.value);
+    }
+  }
 
   Future<void> animateToStep(int index) async {
     final double stepAngle = (2 * math.pi) - (index * _stepAngle);
@@ -202,20 +248,20 @@ class _CarouselWidget3DState extends State<CarouselWidget3D>
   }
 
   Future<void> animateToClosestStep() async {
-    final int closestStep = (_controller.value / _stepAngle).round();
+    final int closestStep = (thetaController.value / _stepAngle).round();
     final double closestStepAngle = closestStep * _stepAngle;
     await animateTo(closestStepAngle, widget.snapTimeInMillis);
   }
 
   Future<void> animateToPreviousStep() async {
-    final int currentStep = (_controller.value / _stepAngle).round();
+    final int currentStep = (thetaController.value / _stepAngle).round();
     final int nextStep = currentStep - 1;
     final double nextAngle = nextStep * _stepAngle;
     await animateTo(nextAngle);
   }
 
   Future<void> animateToNextStep() async {
-    final int currentStep = (_controller.value / _stepAngle).round();
+    final int currentStep = (thetaController.value / _stepAngle).round();
     final int nextStep = currentStep + 1;
     final double nextAngle = nextStep * _stepAngle;
     await animateTo(nextAngle);
@@ -227,35 +273,35 @@ class _CarouselWidget3DState extends State<CarouselWidget3D>
   ]) async {
     // If the direct linear path between _controller.value and nextAngle
     // is shorter than half the circle (pi radians), directly animate to nextAngle
-    if ((_controller.value - nextAngle).abs() < math.pi) {
-      await _controller.animateTo(
+    if ((thetaController.value - nextAngle).abs() < math.pi) {
+      await thetaController.animateTo(
         nextAngle,
         duration: Duration(milliseconds: durationInMillis),
       );
     } else {
       late final double duration1;
-      if (_controller.value <= math.pi) {
+      if (thetaController.value <= math.pi) {
         duration1 = durationInMillis *
-            ((_controller.value - _controller.lowerBound) /
-                ((_controller.value - _controller.lowerBound) +
-                    (_controller.upperBound - nextAngle)));
-        await _controller.animateTo(
-          _controller.lowerBound,
+            ((thetaController.value - thetaController.lowerBound) /
+                ((thetaController.value - thetaController.lowerBound) +
+                    (thetaController.upperBound - nextAngle)));
+        await thetaController.animateTo(
+          thetaController.lowerBound,
           duration: Duration(milliseconds: duration1.round()),
         );
-        _controller.value = _controller.upperBound;
+        thetaController.value = thetaController.upperBound;
       } else {
         duration1 = durationInMillis *
-            ((_controller.upperBound - _controller.value) /
-                ((_controller.upperBound - _controller.value) +
-                    (nextAngle - _controller.lowerBound)));
-        await _controller.animateTo(
-          _controller.upperBound,
+            ((thetaController.upperBound - thetaController.value) /
+                ((thetaController.upperBound - thetaController.value) +
+                    (nextAngle - thetaController.lowerBound)));
+        await thetaController.animateTo(
+          thetaController.upperBound,
           duration: Duration(milliseconds: duration1.round()),
         );
-        _controller.value = _controller.lowerBound;
+        thetaController.value = thetaController.lowerBound;
       }
-      await _controller.animateTo(
+      await thetaController.animateTo(
         nextAngle,
         duration: Duration(milliseconds: durationInMillis - duration1.round()),
       );
@@ -268,46 +314,76 @@ class _CarouselWidget3DState extends State<CarouselWidget3D>
     return (dx / radius) * widget.dragSensitivity;
   }
 
-  double initialPosition = 0;
-
-  double getOffsetAngle(int index) =>
-      (_controller.value + (index * _stepAngle)) % (2 * math.pi);
-
-  double getXTranslation(int index) =>
-      widget.radius *
-      math.sin(getOffsetAngle(index)) *
-      (widget.clockwise ? -1 : 1);
-
-  double getZTranslation(int index) =>
-      widget.radius * math.cos(getOffsetAngle(index));
-
-  double getYRotation(int index) =>
-      widget.spinWhileRotating ? -getOffsetAngle(index) : 0;
-
-  void onDragStart(details) {
-    final bool isHorizontal = widget.spinAxis == Axis.horizontal;
-
-    initialPosition =
-        isHorizontal ? details.globalPosition.dx : details.globalPosition.dy;
+  double normalizeAlphaForYCalculations(double valueInRadians) {
+    valueInRadians = valueInRadians % (math.pi * 2);
+    if (valueInRadians >= 0 && valueInRadians <= math.pi / 2) {
+      return valueInRadians;
+    } else if (valueInRadians <= math.pi) {
+      return math.pi - valueInRadians;
+    } else if (valueInRadians <= (3 * math.pi / 2)) {
+      return math.pi - valueInRadians;
+    } else {
+      return valueInRadians - (2 * math.pi);
+    }
   }
 
-  void onDragUpdate(details) {
-    final bool isHorizontal = widget.spinAxis == Axis.horizontal;
+  double initialPosition = 0;
+
+  double getThetaOffset(int index) =>
+      (thetaController.value + (index * _stepAngle)) % (2 * math.pi);
+
+  double getXTranslation(int index) =>
+      widget.radius * math.sin(getThetaOffset(index));
+  // (widget.clockwise ? -1 : 1); // TODO : Consider clockwise-ness
+
+  double getYTranslation(int index) =>
+      widget.radius *
+      math.sin(alphaController.value) *
+      math.cos(getThetaOffset(index));
+
+  double getZTranslation(int index) =>
+      widget.radius -
+      (widget.radius *
+          math.cos(getThetaOffset(index)) *
+          math.cos(alphaController.value));
+
+  double getYRotation(int index) =>
+      widget.childrenAreAlwaysFacingForward ? 0 : -getThetaOffset(index);
+
+  double getXRotation(int index) =>
+      widget.childrenAreAlwaysFacingForward ? 0 : 0;
+
+  double getZRotation(int index) =>
+      widget.childrenAreAlwaysFacingForward ? 0 : 0;
+
+  void onDragStart(DragStartDetails details) {
+    final bool isDragHorizontal = widget.dragDirection == Axis.horizontal;
+
+    initialPosition = isDragHorizontal
+        ? details.globalPosition.dx
+        : details.globalPosition.dy;
+  }
+
+  void onDragUpdate(DragUpdateDetails details) {
+    final bool isDragHorizontal = widget.dragDirection == Axis.horizontal;
 
     final double dragDistance = initialPosition -
-        (isHorizontal ? details.globalPosition.dx : details.globalPosition.dy);
-    initialPosition =
-        isHorizontal ? details.globalPosition.dx : details.globalPosition.dy;
+        (isDragHorizontal
+            ? details.globalPosition.dx
+            : details.globalPosition.dy);
+    initialPosition = isDragHorizontal
+        ? details.globalPosition.dx
+        : details.globalPosition.dy;
     final double angleInRadian = widget.clockwise
         ? -_getAngleInRadFromXMovement(dragDistance)
         : _getAngleInRadFromXMovement(dragDistance);
     animateTo(
-      (_controller.value - angleInRadian) % (2 * math.pi),
+      (thetaController.value - angleInRadian) % (2 * math.pi),
       0,
     );
   }
 
-  void onDragEnd(details) {
+  void onDragEnd(DragEndDetails details) {
     switch (widget.dragEndBehavior) {
       case DragEndBehavior.doNothing:
         break;
@@ -315,8 +391,8 @@ class _CarouselWidget3DState extends State<CarouselWidget3D>
         animateToClosestStep();
         break;
       case DragEndBehavior.continueRotating:
-        if (widget.shouldRotate) {
-          _rotateInfinitely();
+        if (widget.thetaShouldRotate) {
+          thetaController.repeat();
         } else {
           animateToClosestStep();
         }
@@ -326,29 +402,29 @@ class _CarouselWidget3DState extends State<CarouselWidget3D>
 
   @override
   Widget build(BuildContext context) {
-    final bool isHorizontal = widget.spinAxis == Axis.horizontal;
+    final bool isDragHorizontal = widget.dragDirection == Axis.horizontal;
 
     return GestureDetector(
       onTap: _onBackgroundTap,
       onHorizontalDragStart:
-          widget.isDragInteractive && isHorizontal ? onDragStart : null,
+          widget.isDragInteractive && isDragHorizontal ? onDragStart : null,
       onHorizontalDragUpdate:
-          widget.isDragInteractive && isHorizontal ? onDragUpdate : null,
+          widget.isDragInteractive && isDragHorizontal ? onDragUpdate : null,
       onHorizontalDragEnd:
-          widget.isDragInteractive && isHorizontal ? onDragEnd : null,
+          widget.isDragInteractive && isDragHorizontal ? onDragEnd : null,
       onVerticalDragStart:
-          widget.isDragInteractive && !isHorizontal ? onDragStart : null,
+          widget.isDragInteractive && !isDragHorizontal ? onDragStart : null,
       onVerticalDragUpdate:
-          widget.isDragInteractive && !isHorizontal ? onDragUpdate : null,
+          widget.isDragInteractive && !isDragHorizontal ? onDragUpdate : null,
       onVerticalDragEnd:
-          widget.isDragInteractive && !isHorizontal ? onDragEnd : null,
+          widget.isDragInteractive && !isDragHorizontal ? onDragEnd : null,
       child: AnimatedBuilder(
-        animation: _controller,
+        animation: thetaController,
         builder: (context, child) {
           List<(int, double)> indexToDistFrom180 = [];
           for (int i = 0; i < widget.children.length; i++) {
             final double offsetAngle = UtilFunctions.inDegrees(
-              getOffsetAngle(i),
+              getThetaOffset(i),
             );
             indexToDistFrom180.add((i, (180 - offsetAngle).abs()));
           }
@@ -378,14 +454,15 @@ class _CarouselWidget3DState extends State<CarouselWidget3D>
                     for (int i = 0; i < widget.children.length; i++)
                       if (indexToDistFrom180[i].$2 < 90)
                         _SubWidget(
-                          isHorizontal: isHorizontal,
                           scale: widget.childScale,
                           radius: widget.radius,
                           perspectiveStrength: widget.perspectiveStrength,
                           xTranslation: getXTranslation(getIndexOf(i)),
-                          zTranslation:
-                              widget.radius - getZTranslation(getIndexOf(i)),
+                          yTranslation: getYTranslation(getIndexOf(i)),
+                          zTranslation: getZTranslation(getIndexOf(i)),
                           yRotation: getYRotation(getIndexOf(i)),
+                          xRotation: getXRotation(getIndexOf(i)),
+                          zRotation: getZRotation(getIndexOf(i)),
                           child: widget.children[getIndexOf(i)].child,
                           onTap: () {
                             switch (widget.childTapBehavior) {
@@ -424,19 +501,36 @@ class _CarouselWidget3DState extends State<CarouselWidget3D>
                       child: widget.core!,
                     ),
 
+                  //TODO:
+                  Column(
+                    mainAxisSize: MainAxisSize.min,
+                    mainAxisAlignment: MainAxisAlignment.center,
+                    crossAxisAlignment: CrossAxisAlignment.center,
+                    children: [
+                      Text(
+                        (thetaController.value * 180 / math.pi)
+                            .toStringAsFixed(0),
+                      ),
+                      Text(
+                        (getZTranslation(getIndexOf(0))).toStringAsFixed(0),
+                      ),
+                    ],
+                  ),
+
                   // Front half of the carousel
                   for (int i = 0; i < widget.children.length; i++)
                     if (indexToDistFrom180[i].$2 >= 90)
                       _SubWidget(
-                        isHorizontal: isHorizontal,
                         scale: widget.childScale,
                         radius: widget.radius,
                         perspectiveStrength: widget.perspectiveStrength,
-                        xTranslation: getXTranslation(indexToDistFrom180[i].$1),
-                        zTranslation: widget.radius -
-                            getZTranslation(indexToDistFrom180[i].$1),
-                        yRotation: getYRotation(indexToDistFrom180[i].$1),
-                        child: widget.children[indexToDistFrom180[i].$1].child,
+                        xTranslation: getXTranslation(getIndexOf(i)),
+                        yTranslation: getYTranslation(getIndexOf(i)),
+                        zTranslation: getZTranslation(getIndexOf(i)),
+                        yRotation: getYRotation(getIndexOf(i)),
+                        xRotation: getXRotation(getIndexOf(i)),
+                        zRotation: getZRotation(getIndexOf(i)),
+                        child: widget.children[getIndexOf(i)].child,
                         onTap: () {
                           switch (widget.childTapBehavior) {
                             case ChildTapBehavior.doNothing:
@@ -465,17 +559,17 @@ class _CarouselWidget3DState extends State<CarouselWidget3D>
   void _onBackgroundTap() {
     switch (widget.backgroundTapBehavior) {
       case BackgroundTapBehavior.startAndSnapToNearest:
-        if (!isAnimating) {
-          _rotateInfinitely();
+        if (!thetaIsAnimating) {
+          thetaController.repeat();
         } else {
           animateToClosestStep();
         }
         break;
       case BackgroundTapBehavior.startAndFreeze:
-        if (!isAnimating) {
-          _rotateInfinitely();
+        if (!thetaIsAnimating) {
+          thetaController.repeat();
         } else {
-          _controller.value = _controller.value;
+          thetaController.value = thetaController.value;
         }
         break;
       case BackgroundTapBehavior.none: