+++ lib/src/models/models.dart
@@ -2,3 +2,4 @@ export 'background_tap_behavior.dart';
 export 'child_tap_behavior.dart';
 export 'carousel_child.dart';
 export 'drag_behaviour.dart';
+export 'tilt_controller.dart';