+++ lib/src/models/tilt_controller.dart
@@ -1,0 +1,11 @@
+class TiltController {
+  final double initialTiltAngle;
+  final double axisOffset;
+  final int tiltAnimationTime;
+
+  const TiltController({
+    this.initialTiltAngle = 0,
+    this.axisOffset = 0,
+    this.tiltAnimationTime = 10000,
+  });
+}